renamed and created the molecule for input
