+++ src/atoms/InputField/index.ts
@@ -1,2 +1,2 @@
-export { default as Input } from "./Input";
-export * from "./Input.model";
+export { default as Inputfield } from "./Inputfield";
+export * from "./Inputfield.model";
