+++ src/atoms/index.ts
@@ -1,3 +1,3 @@
 export { Button } from "./Button";
-export { Input } from "./InputField";
+export { Inputfield } from "./InputField";
 export { Label } from "./InputLabel";
