+++ src/molecules/Input/Input.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { Inputfield, Label } from "../../atoms";
+import { ICustomInput } from "../../atoms/InputField";
+
+const Input = ({
+  ...props
+}: React.PropsWithChildren<ICustomInput>): React.ReactElement => {
+  return (
+    <div>
+      <Label />
+      <Inputfield {...props} />
+    </div>
+  );
+};
+
+export default Input;
